fix: colocate Petri transition assertions in assertions.go for pkg file-count
Merge petriassert helpers into assertions.go so pkg/testutil stays within
the 15-file package limit; CI make pkg-file-count was failing on PR #518.
